requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
Flask==3.0.2
spotipy==2.23.0
urllib3==1.26.6
python-dotenv==1.0.1